refactor: ReservHistoryAdapter 패키지 변경(#152)
