+++ presentation/src/main/java/com/example/smiley/main/reserv/adapter/ReservHistoryAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.smiley.main.reserv
+package com.example.smiley.main.reserv.adapter
 
 import android.annotation.SuppressLint
 import android.content.res.ColorStateList
